Modify Readme and deploy on Rinkeby
The smart contracts have been deployed on Rinkeby,
the configuration for this has been added to the truffle config file.
The README file has been modified to give infos about the RInkeby deployment.
In the README file an explanation of how to test the smart contracts on a local
server has been added.